Fix button layout
